requirements.txt
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
kaggle==1.7.4.5
numpy==2.2.6
pandas==2.3.2
protobuf==6.32.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
requests==2.32.5
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
six==1.17.0
SQLAlchemy==2.0.43
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
webencodings==0.5.1
streamlit>=1.28.0
plotly>=5.17.0